refactor to support interface for non-block requests
